pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "orchestra-agents"
version = "0.1.0"
description = "Python-first multi-agent orchestration framework"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Orchestra Contributors" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    "pydantic>=2.5",
    "httpx>=0.26",
    "structlog>=24.0",
    "rich>=13.0",
    "typer>=0.12",
    "mcp>=1.26,<2",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.20"]
google = ["google-generativeai>=0.5"]
ollama = []  # No SDK needed, uses httpx (already in core)
all-providers = ["orchestra-agents[anthropic,google,nats,ray]"]
storage = ["aiosqlite>=0.19"]
postgres = ["asyncpg>=0.29"]
telemetry = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-http>=1.20",
    "opentelemetry-semantic-conventions>=0.45b0",
]
cache = [
    "cachetools>=5.5",
    "diskcache>=5.6",
    "ruamel.yaml>=0.18",
]
server = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sse-starlette>=2.1",
    "pydantic-settings>=2.0",
    "watchfiles>=0.21",
]
routing = ["numpy>=1.26"]
security = [
    "wasmtime>=23.0",
    "joserfc>=1.6",
]
redis = ["redis[hiredis]>=7.0", "msgpack>=1.0"]
memory = [
    "redis[hiredis]>=7.0",
    "msgpack>=1.0",
]
vectordb = [
    "pgvector>=0.4",
    "model2vec>=0.6",
    "pyzstd>=0.17",
    "msgpack>=1.0",
]
qdrant = [
    "qdrant-client>=1.14",
]
promptshield = [
    "optimum[onnxruntime]>=1.24",
    "transformers>=4.51,<5.0",
]
nats = ["nats-py>=2.14"]
messaging = [
    "nats-py>=2.14",
    "joserfc>=1.6",
    "base58>=2.1",
    "cryptography>=42.0",
    "aiohttp>=3.9",
]
ray = ["ray[serve]>=2.10"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
    "pre-commit>=3.6",
]
test-advanced = [
    "locust>=2.20",
    "hypothesis>=6.100",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.24",
]

[project.scripts]
orchestra = "orchestra.cli.main:app"

[project.urls]
Homepage = "https://github.com/orchestra-agents/orchestra"
Documentation = "https://orchestra-agents.dev"
Repository = "https://github.com/orchestra-agents/orchestra"

[tool.hatch.build.targets.wheel]
packages = ["src/orchestra"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
    "load: marks load/stress tests requiring 1000+ concurrent tasks (deselect with '-m \"not load\"')",
    "live: marks tests that hit real LLM APIs — requires credentials or Ollama (deselect with '-m \"not live\"')",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W",    # pycodestyle
    "F",         # pyflakes
    "I",         # isort
    "UP",        # pyupgrade
    "B",         # bugbear
    "SIM",       # simplify
    "RUF",       # ruff-specific
]
ignore = [
    "UP042",     # keep str enum pattern for Python 3.11 compat
    "SIM105",    # contextlib.suppress breaks when try assigns a variable
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["RUF012", "B017", "S101"]
"examples/**" = ["RUF012"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
source = ["orchestra"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
[tool.bandit]
# B104: binding to 0.0.0.0 is intentional for an HTTP server
# B608: SQL f-strings in vector_store.py use parameterized $N values for user data;
#       only table_name (a trusted constructor arg) is interpolated directly
skips = ["B104", "B608"]
